partial navigate()
