+++ src/runtime/entrypoints/impl/global.ts
@@ -6,6 +6,7 @@ export type {
   IslandRegistry,
   RenderRequest,
 } from './common.ts';
+export type { FreshHistoryState } from './partials.ts'
 
 function getGlobal(name: string) {
   return (window as any)[GLOBAL_SYMBOL_PREFIX + name]
@@ -13,4 +14,6 @@ function getGlobal(name: string) {
 
 export const _walkInner = getGlobal('_walkInner') as typeof commonImpl._walkInner
 export const NoPartialsError = getGlobal('NoPartialsError') as typeof partialImpl.NoPartialsError
+export const historyState = getGlobal('historyState')
+export const navigate = getGlobal('navigate') as typeof partialImpl.navigate
 export const fetchPartials = getGlobal('fetchPartials') as typeof partialImpl.fetchPartials
