+++ src/runtime/entrypoints/impl/globalInit.ts
@@ -1,6 +1,6 @@
 import { GLOBAL_SYMBOL_PREFIX } from "$fresh/src/constants.ts"
 import { _walkInner, isCommentNode, isElementNode, isTextNode } from "./common.ts"
-import { NoPartialsError, PartialComp, applyPartials, fetchPartials } from "./partials.ts"
+import { NoPartialsError, PartialComp, applyPartials, fetchPartials, historyState, navigate } from "./partials.ts"
 
 function assignGlobal(name: string, obj: any) {
   (window as any)[GLOBAL_SYMBOL_PREFIX + name] = obj
@@ -14,3 +14,5 @@ assignGlobal('NoPartialsError', NoPartialsError)
 assignGlobal('PartialComp', PartialComp)
 assignGlobal('applyPartials', applyPartials)
 assignGlobal('fetchPartials', fetchPartials)
+assignGlobal('historyState', historyState)
+assignGlobal('navigate', navigate)
